+++ projects/iworb/rxjs-utils/src/lib/index.ts
@@ -1,1 +1,3 @@
 export * from './combineLatestMap';
+export * from './downloadWaterfall';
+export * from './untilDestroyedService';