+++ projects/iworb/rxjs-utils/src/public-api.ts
@@ -2,4 +2,4 @@
  * Public API Surface of rxjs-utils
  */
 
-export * from './lib/combineLatestMap';
+export * from './lib/index';