+++ src/app/app.module.ts
@@ -1,3 +1,4 @@
+import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
@@ -10,18 +11,25 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgxJsonViewerModule } from 'ngx-json-viewer';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { SidenavService } from './services';
 import { CombineLatestMapExampleComponent } from './combine-latest-map-example/combine-latest-map-example.component';
+import { DownloadExampleComponent } from './download-example/download-example.component';
+import { SidenavService } from './services';
+import { UntilDestroyedExampleComponent } from './until-destroyed-example/until-destroyed-example.component';
+import { UntilDestroyedComponent } from './until-destroyed-example/until-destroyed/until-destroyed.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    CombineLatestMapExampleComponent
+    CombineLatestMapExampleComponent,
+    UntilDestroyedExampleComponent,
+    UntilDestroyedComponent,
+    DownloadExampleComponent
   ],
   imports: [
+    AppRoutingModule,
     BrowserModule,
     BrowserAnimationsModule,
-    AppRoutingModule,
+    HttpClientModule,
 
     MatButtonModule,
     MatCardModule,