+++ src/app/until-destroyed-example/until-destroyed/until-destroyed.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { UntilDestroyedComponent } from './until-destroyed.component';
+
+describe('UntilDestroyedComponent', () => {
+  let component: UntilDestroyedComponent;
+  let fixture: ComponentFixture<UntilDestroyedComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ UntilDestroyedComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(UntilDestroyedComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});